plant-disease-detection-app/backend/requirements.txt
tensorflow
Pillow
numpy
Flask
Flask-CORS
h5py
werkzeug